Add documentation for TextMenus
Add more logic when merging documentation

+++ Celeste.Mod.mm/Mod/UI/TextMenuExt.cs
@@ -236,7 +236,7 @@ namespace Celeste {
         }
 
         /// <summary>
-        /// Convenience class for creating a <see cref="TextMenu.Option{T}"/> from an enum class.<br></br>
+        /// Convenience class for creating a <see cref="TextMenu.Option{T}"/> from an <see cref="Enum"/>.<br></br>
         /// Not to be confused with <see cref="EnumerableSlider{T}"/>
         /// </summary>
         /// <typeparam name="T">Enum Type</typeparam>
@@ -298,6 +298,7 @@ namespace Celeste {
             public int Index;
             public Action<int> OnValueChange;
             public int PreviousIndex;
+
             private float sine;
             private int lastDir;
             private int min;
@@ -318,13 +319,17 @@ namespace Celeste {
                 this.max = max;
                 Index = (value < min) ? min : (value > max) ? max : value;
             }
+
+            /// <inheritdoc cref="TextMenu.Option{T}.Change(Action{T})"/>
             public IntSlider Change(Action<int> action) {
                 OnValueChange = action;
                 return this;
             }
+
             public override void Added() {
                 Container.InnerContent = TextMenu.InnerContentMode.TwoColumn;
             }
+
             public override void LeftPressed() {
                 if (Input.MenuLeft.Repeating)
                     fastMoveTimer += Engine.DeltaTime * 8;
@@ -341,6 +346,7 @@ namespace Celeste {
                     OnValueChange?.Invoke(Index);
                 }
             }
+
             public override void RightPressed() {
                 if (Input.MenuRight.Repeating)
                     fastMoveTimer += Engine.DeltaTime * 8;
@@ -357,6 +363,7 @@ namespace Celeste {
                     OnValueChange?.Invoke(Index);
                 }
             }
+
             public override void ConfirmPressed() {
                 if ((max - min) == 1) {
                     if (Index == min) {
@@ -375,14 +382,17 @@ namespace Celeste {
             public override void Update() {
                 sine += Engine.RawDeltaTime;
             }
+
             public override float LeftWidth() {
                 return ActiveFont.Measure(Label).X + 32f;
             }
+
             public override float RightWidth() {
                 // Measure Index in case it is externally set ouside the bounds
                 float width = Calc.Max(0f, ActiveFont.Measure(max.ToString()).X, ActiveFont.Measure(min.ToString()).X, ActiveFont.Measure(Index.ToString()).X);
                 return width + 120f;
             }
+
             public override float Height() {
                 return ActiveFont.LineHeight;
             }
@@ -409,19 +419,22 @@ namespace Celeste {
 
         /// <summary>
         /// <see cref="TextMenu.Item"/> that acts as a Submenu for other Items.
-        /// <br></br><br></br>
+        /// <br/><br/>
         /// Currently does not support recursive submenus
         /// </summary>
         public class SubMenu : TextMenu.Item {
             public string Label;
             MTexture Icon;
 
+            /// <inheritdoc cref="patch_TextMenu.Items"/>
             public List<TextMenu.Item> Items { get; private set; }
 
             private List<TextMenu.Item> delayedAddItems;
 
+            /// <inheritdoc cref="TextMenu.Selection"/>
             public int Selection;
 
+            /// <inheritdoc cref="TextMenu.Current"/>
             public TextMenu.Item Current {
                 get {
                     if (Items.Count <= 0 || Selection < 0) {
@@ -434,6 +447,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.FirstPossibleSelection"/>
             public int FirstPossibleSelection {
                 get {
                     for (int i = 0; i < Items.Count; i++) {
@@ -445,6 +459,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.LastPossibleSelection"/>
             public int LastPossibleSelection {
                 get {
                     for (int i = Items.Count - 1; i >= 0; i--) {
@@ -456,18 +471,22 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.ScrollTargetY"/>
             public float ScrollTargetY {
                 get {
                     float min = Engine.Height - 150f - Container.Height * Container.Justify.Y;
                     float max = 150f + Container.Height * Container.Justify.Y;
-                    return Calc.Clamp((float) (Engine.Height / 2) + Container.Height * Container.Justify.Y - GetYOffsetOf(Current), min, max);
+                    return Calc.Clamp((Engine.Height / 2) + Container.Height * Container.Justify.Y - GetYOffsetOf(Current), min, max);
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.ItemSpacing"/>
             public float ItemSpacing;
             public float ItemIndent;
+            /// <inheritdoc cref="TextMenu.HighlightColor"/>
             private Color HighlightColor;
             public string ConfirmSfx;
+
             public bool AlwaysCenter;
 
             public float LeftColumnWidth;
@@ -484,7 +503,7 @@ namespace Celeste {
             private bool containerAutoScroll;
 
             /// <summary>
-            /// 
+            /// Create a new SubMenu.
             /// </summary>
             /// <param name="label"></param>
             /// <param name="enterOnSelect">Expand submenu when selected</param>
@@ -588,24 +607,31 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.Clear"/>
             public void Clear() {
                 Items = new List<TextMenu.Item>();
             }
 
+            /// <inheritdoc cref="TextMenu.IndexOf(TextMenu.Item)"/>
             public int IndexOf(TextMenu.Item item) {
                 return Items.IndexOf(item);
             }
 
+            /// <inheritdoc cref="TextMenu.FirstSelection"/>
             public void FirstSelection() {
                 Selection = -1;
                 MoveSelection(1, false);
             }
 
+            /// <summary>
+            /// Set the selection to the last possible <see cref="TextMenu.Item"/>.
+            /// </summary>
             public void LastSelection() {
                 Selection = LastPossibleSelection;
                 MoveSelection(0, false);
             }
 
+            /// <inheritdoc cref="TextMenu.MoveSelection(int, bool)"/>
             public void MoveSelection(int direction, bool wiggle = false) {
                 int selection = Selection;
                 direction = Math.Sign(direction);
@@ -653,6 +679,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.RecalculateSize"/>
             public void RecalculateSize() {
                 TitleHeight = ActiveFont.LineHeight;
                 if (Items.Count < 1)
@@ -677,6 +704,7 @@ namespace Celeste {
                 MenuHeight -= Container.ItemSpacing;
             }
 
+            /// <inheritdoc cref="TextMenu.GetYOffsetOf(TextMenu.Item)"/>
             public float GetYOffsetOf(TextMenu.Item item) {
                 if (item == null) {
                     return 0f;
@@ -850,7 +878,7 @@ namespace Celeste {
 
         /// <summary>
 		/// <see cref="TextMenu.Item"/> that acts as a Slider of Submenus for other Items.
-		/// <br></br><br></br>
+		/// <br/><br/>
 		/// Currently does not support recursive submenus
 		/// </summary>
 		public class OptionSubMenu : TextMenu.Item {
@@ -865,17 +893,25 @@ namespace Celeste {
             public int MenuIndex;
 
             private int InitialSelection;
+            ///<inheritdoc cref="TextMenu.Selection"/>
             public int Selection;
 
             private int lastDir;
             private float sine;
 
+            /// <summary>
+            /// Invoked when the selected menu is changed.
+            /// </summary>
             public Action<int> OnValueChange;
 
+            /// <summary>
+            /// The selected set of <see cref="TextMenu.Item"/>s.
+            /// </summary>
             public List<TextMenu.Item> CurrentMenu {
                 get { return (Menus.Count > 0) ? Menus[MenuIndex].Item2 : null; }
             }
 
+            /// <inheritdoc cref="TextMenu.Current"/>
             public TextMenu.Item Current {
                 get {
                     if (CurrentMenu.Count <= 0 || Selection < 0) {
@@ -885,6 +921,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.FirstPossibleSelection"/>
             public int FirstPossibleSelection {
                 get {
                     for (int i = 0; i < CurrentMenu.Count; i++) {
@@ -896,6 +933,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.LastPossibleSelection"/>
             public int LastPossibleSelection {
                 get {
                     for (int i = CurrentMenu.Count - 1; i >= 0; i--) {
@@ -907,6 +945,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.ScrollTargetY"/>
             public float ScrollTargetY {
                 get {
                     float min = Engine.Height - 150 - Container.Height * Container.Justify.Y;
@@ -915,6 +954,7 @@ namespace Celeste {
                 }
             }
 
+            /// <inheritdoc cref="TextMenu.ItemSpacing"/>
             public float ItemSpacing;
             public float ItemIndent;
 
@@ -1004,12 +1044,14 @@ namespace Celeste {
                 Menus = new List<Tuple<string, List<TextMenu.Item>>>();
             }
 
+            /// <inheritdoc cref="TextMenu.FirstSelection"/>
             public void FirstSelection() {
                 Selection = -1;
                 if (CurrentMenu.Count > 0)
                     MoveSelection(1, true);
             }
 
+            /// <inheritdoc cref="TextMenu.MoveSelection(int, bool)"/>
             public void MoveSelection(int direction, bool wiggle = false) {
                 int selection = Selection;
                 direction = Math.Sign(direction);
@@ -1094,6 +1136,9 @@ namespace Celeste {
 
             #region TextMenu.Item
 
+            /// <summary>
+            /// Set the action to be invoked when the selected menu is changed.
+            /// </summary>
             public OptionSubMenu Change(Action<int> onValueChange) {
                 OnValueChange = onValueChange;
                 return this;
